# src/main/resources/openApi-Team.yml
openapi: 3.0.0
info:
  title: Team
  version: 1.0.0
paths:
  /team/{teamId}:
    get:
      operationId: retrieve
      summary: Returns your team players
      description: Returns an ordered list based on your player points.
      tags:
        - Team
      parameters:
        - in: path
          name: teamId
          schema:
            type: integer
          required: true
          description: ID of the team
      responses:
        '200':
          description: Team found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDto'
        '500':
          description: Error during team retrieval
  /team:
    post:
      operationId: create
      summary: Create a team.
      description: Create an empty team.
      tags:
        - Team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamToCreateDto'
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamCreatedDto'
        '400':
          description: Wrong input body
        '500':
          description: Error during team retrieval

components:
  schemas:
    TeamDto:
      type: object
      properties:
        players:
          $ref: '#/components/schemas/TeamPlayersDto'
        totalScore:
          type: double
          default: 0
          description: Total team's score
      required:
        - players
        - totalScore

    TeamPlayersDto:
      type: array
      items:
        $ref: '#/components/schemas/TeamPlayerDto'

    TeamPlayerDto:
      type: object
      properties:
        fullName:
          type: string
          default: ""
          description: Player's full name
        fantaPoints:
          type: double
          default: 0
          description: Player's fanta-points
      required:
        - fullName
        - fantaPoints

    TeamCreatedDto:
      type: object
      properties:
        id:
          type: integer
          default: 0
          description: Team ID
        ownerId:
          type: string
          default: ""
          description: Team's Owner ID
      required:
        - id
        - ownerId

    TeamToCreateDto:
      type: object
      properties:
        ownerId:
          type: string
          default: ""
          example: OWNER_ID
      required:
        - ownerId